Add Catilin, Doors and refactor Cutscene
