+++ data/quest/free/blood_pact/blood_pact.anims.toml
@@ -1,4 +1,3 @@
-# all anim ids wrong
 ["kayle_defeat"]
 id = 12851
 
@@ -7,3 +6,18 @@ id = 12880
 
 ["kayle_getUp"]
 id = 12856
+
+["caitlin_defeat"]
+id = 12851
+
+["caitlin_death"]
+id = 12880
+
+["caitlin_getUp"]
+id = 12856
+
+["reese_defeat"]
+id = 12851
+
+["reese_death"]
+id = 12880
