+++ data/quest/free/blood_pact/blood_pact.teles.toml
