+++ data/quest/free/blood_pact/blood_pact.vars.toml
@@ -1,0 +1,17 @@
+[blood_pact_kayle]
+format = "list"
+persist = true
+default = "alive"
+values = ["alive", "defeated", "spared", "killed"]
+
+[blood_pact_caitlin]
+format = "list"
+persist = true
+default = "alive"
+values = ["alive", "defeated", "spared", "killed"]
+
+[blood_pact_reese]
+format = "list"
+persist = true
+default = "alive"
+values = ["alive", "defeated", "spared", "killed"]
